server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# server/config.py: 2
# server/db.py: 7
PyYAML == 5.3.1

# server/api.py: 1
# server/web.py: 2
Quart == 0.13.0

# server/api.py: 5
Quart_CORS == 0.3.0

# server/db.py: 3,4
asyncpg == 0.21.0

# server/db.py: 11
python_dateutil == 2.8.1

# server/fetcher.py: 11
uvloop == 0.14.0

# server/fetcher.py: 4
websockets == 8.1

# server/fetcher.py: 6,7
# server/web.py: 7,8
zmq == 0.0.0
